Use a compact representation of accumulator trees
Since I modified the use of accumulator trees so that information
is only stored on right children, not on the root or left children,
it didn't make sense to represent the entire tree in a vector anymore.
By eliminating half + 1 of the items in the tree, we can still consistently
access right children, while pretending that the left children exist.