Refine the making of FlatGraphs
It turned out that we needed both Edges and Segments to be able to
create blocks later in the process, so I reworked the FlatGraph
code to include them. This removed the need to transform :preds
and :succs into specifically flat versions. I also brought :marked
along for the ride; not including it originally was an oversight.