Refactor plotting y coordinates
I was never happy with YPlottedClassGraph, so I removed it. Instead,
I merged its functionality into a regular ClassGraph, and rather
than storing a :ys key, I defer to a memoized ys function. In the
process, I also split up the work differently between functions,
and I'm much happier with the way this part of the code feels now.
Also, I got to name a function memo-ys. Hard to beat that.