Refactor plotting
The previous iteration was complicated and buggy; this one is
hopefully less so. The biggest conceptual change is how ClassGraphs
are created and processed. FlatGraphs are no longer flipped and
reversed to generate four derivative FlatGraphs that are each
processed into ClassGraphs. Instead, the FlatGraph is reversed, and
block info is generated from both variations. Then, each block-info
is compacted both up and down into two ClassGraphs, for a total of
four ClassGraphs. Plotting y-values from these ClassGraphs works
similarly as before.

There was also a lighter refactoring of the y-plotting. Factoring
the core logic into a higher order function wasn't as much of a
win as I'd hoped, and it made the process pretty difficult to
understand. It's better now.