Tweak arc radius calc
It turned out that in some cases, ordering based on contiguous
groups of :up and :down per layer in the way we were still causes
some arcs to get too close to each other.  This tweak calculates
the radii based on a total ordering of :up and :down in a layer.